Separate out webdriver test and screenshot test

+++ run_test.sh
@@ -63,13 +63,29 @@ function run_webdriver_test {
   export FLASK_ENV=WEBDRIVER
   export GOOGLE_CLOUD_PROJECT=datcom-browser-staging
   pip3 install -r requirements.txt -q
+  python3 -m pytest --ignore=webdriver_tests/screenshot_test.py webdriver_tests/**.py
+  cd ..
+}
+
+function run_screenshot_test {
+  python3 -m venv .env
+  source .env/bin/activate
+  cd server
+  if [ ! -d dist  ]
+  then
+    echo "no dist folder, please run ./run_test.sh -b to build js first."
+    exit 1
+  fi
+  export FLASK_ENV=WEBDRIVER
+  export GOOGLE_CLOUD_PROJECT=datcom-browser-staging
+  pip3 install -r requirements.txt -q
   if [  -d test_screenshots  ]
   then
     echo "Delete the test_screenshots folder"
     rm -rf test_screenshots
   fi
   mkdir test_screenshots
-  python3 -m pytest webdriver_tests/**.py
+  python3 -m pytest webdriver_tests/screenshot_test.py
   cd ..
 }
 
@@ -82,12 +98,13 @@ function run_all_tests {
 }
 
 function help {
-  echo "Usage: $0 -pwblca"
+  echo "Usage: $0 -pwblcas"
   echo "-p       Run server python tests"
   echo "-w       Run webdriver tests"
   echo "-b       Run client install and build"
   echo "-l       Run client lint"
   echo "-c       Run client tests"
+  echo "-s       Run screenshot tests"
   echo "-a       Run all tests"
   echo "No args  Run all tests"
   exit 1
@@ -114,6 +131,10 @@ while getopts pwblca OPTION; do
     c)
         echo -e "### Running client tests"
         run_npm_test
+        ;;
+    s)
+        echo -e "### Running screenshot tests"
+        run_screenshot_test
         ;;
     a)
         echo -e "### Running all tests"
